feat(global): export file created
